sql: dump puts INTERLEAVE clauses outside of CREATE TABLE statements
Fix bug where dump would include INTERLEAVE IN PARENT clauses for
indexes inside of CREATE TABLE statements. CRDB currently does not
support this, thus forcing the user to manually edit the dump files
to remove the INTERLEAVE clauses themselves.

Previousily, crdb_internal.go would use show_create.go to create the
SHOW CREATE TABLE statements which would include INTERLEAVE, for the
create_nofk columns. crdb_internal.go would also add ALTER STATEMENTs
separately to the alter_statements of the create statements of the
table. This change removes the addition of the INTERLEAVE within
show_create.go for the create_nofk columns, and includes the additon
of adding the INTERLEAVE clauses to the alter_statements within
crdb_internal.go.

A bug was also fixed in structured.go in the SQLString method that
created the SQL string out of order when a non-empty table name is
passed in. The bug caused the "ON tablename" to be printed before
the index and without space in between the two.

Resolves: #35462

Release Note (bug fix): dump now works properly when handling
INTERLEAVED tables, printing them outside of CREATE TABLE statements.
