+++ pkg/sql/pg_catalog.go
@@ -817,7 +817,7 @@ CREATE TABLE pg_catalog.pg_constraint (
 						return err
 					}
 					var buf bytes.Buffer
-					if err := printForeignKeyConstraint(ctx, &buf, db.Name, table, con.FK, tableLookup); err != nil {
+					if err := showForeignKeyConstraint(&buf, db.Name, table, con.FK, tableLookup); err != nil {
 						return err
 					}
 					condef = tree.NewDString(buf.String())
