+++ pkg/sql/sqlbase/structured_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
@@ -1999,4 +2000,26 @@ func TestDefaultExprNil(t *testing.T) {
 			}
 		}
 	})
+}
+
+func TestSQLString(t *testing.T) {
+	colNames := []string{"derp", "foo"}
+	indexName := "idx"
+	tableName := tree.MakeTableName("DB", "t1")
+	tableName.ExplicitCatalog = false
+	tableName.ExplicitSchema = false
+	index := IndexDescriptor{Name: indexName,
+		ID:               0x0,
+		Unique:           false,
+		ColumnNames:      colNames,
+		ColumnDirections: []IndexDescriptor_Direction{IndexDescriptor_ASC, IndexDescriptor_ASC},
+	}
+	expected := fmt.Sprintf("INDEX %s ON t1 (%s ASC, %s ASC)", indexName, colNames[0], colNames[1])
+	if got := index.SQLString(&tableName); got != expected {
+		t.Errorf("Expected '%s', but got '%s'", expected, got)
+	}
+	expected = fmt.Sprintf("INDEX %s (%s ASC, %s ASC)", indexName, colNames[0], colNames[1])
+	if got := index.SQLString(&AnonymousTable); got != expected {
+		t.Errorf("Expected '%s', but got '%s'", expected, got)
+	}
 }
